wayfire-ipc.lisp: read-exact: fix for usocket-iolib
* wayfire-ipc.lisp: (read-exact): add an ineffiecent implementaion
conditionalized on usocket-iolib that conses up an array for each
send. for lispworks make the allocation of the response buffer
pinnable.

tldr: iolib does send recv with with-pointer-to-vector-data, which in
lispworks8 uses hcl:with-pinned-objects.

usocket's iolib backend will helpfully (inefficiently) copy data to a
static pinnable buffer when sending but when receiving the onus is on
the receiver to provde a static pinnable buffer.
so we cons up a static vector for every send.

[260707]
* wayfire-ipc.system (*wayfire-ipc-binary-dir*): isolate fasls when
:usocket-iolib is in *features*